Addresses Issue #7
Update evaluates for a hash tag at the end of the URL (currently looks
for #domain-redirected), then fires off an event containing the redirect
as the event category, referral as the event action, and referral source
as event label.  Event non-interaction is set to true so event doesn't
affect the bounce-rate calculation. A message is displayed for the
visitor that the site has moved. closes #7