Fixed #9
Issue was the result of jQuery changes to .attr().

According to jQuery's documentation site, "As of jQuery 1.6, the .attr()
method returns undefined for attributes that have not been set."

We were using:
if ($(this).attr('target').length > 0){}
Changed it to:
if ($(this).attr('target') !== undefined){}

This fix is tested and backwards compatible with 1.4.2-1.9.1.