Closes #8 and closes #10
Realigns 404 error and redirect tracking as events tracked with event
action = destination URL and event label = source URL. Shares function
to clean source URL. Note: 404 error was previously tracked as a
pageview. Moved 404 after redirect tracking (in the event that there's a
redirect to a non-existent page, the destination URL in 404 error event
will not contain the #domain-redirected hash tag (as it's removed in
browser window for redirect tracking).